+++ src/components/modules/epics/ContributionsEpicReminderSignedOut.stories.tsx
@@ -4,7 +4,7 @@ import {
     ContributionsEpicReminderSignedOut,
     ContributionsEpicReminderSignedOutProps,
 } from './ContributionsEpicReminderSignedOut';
-import { ReminderStatus } from './ContributionsEpicReminder';
+import { ReminderStatus } from './utils/reminders';
 import { EpicDecorator } from './ContributionsEpic.stories';
 
 export default {
