+++ src/components/modules/epics/utils/reminders.ts
@@ -1,5 +1,16 @@
 import { addCookie } from '../../../../lib/cookies';
 
+// --- Types --- //
+
+export enum ReminderStatus {
+    Editing = 'Editing',
+    Submitting = 'Submitting',
+    Error = 'Error',
+    Completed = 'Completed',
+}
+
+// --- Cookie utils --- //
+
 const dateDiff = (start: Date, end: Date): number => {
     const twentyFourHours = 86400000;
     return Math.round((end.valueOf() - start.valueOf()) / twentyFourHours);
@@ -12,6 +23,8 @@ export const addContributionReminderCookie = (reminderDateString: string): void 
     addCookie('gu_epic_contribution_reminder', '1', dateDiff(today, reminderDate));
 };
 
+// --- Text utils --- //
+
 const PREPOSITION_REGEX = /^(on|in)/;
 
 const containsPreposition = (text: string): boolean => PREPOSITION_REGEX.test(text);
@@ -20,6 +33,8 @@ const addPreposition = (text: string): string => 'in ' + text;
 
 export const ensureHasPreposition = (text: string): string =>
     containsPreposition(text) ? text : addPreposition(text);
+
+// --- Validation utils --- //
 
 export const isValidEmail = (email: string): boolean => {
     const re = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
